requirements.txt
pytelegrambotapi==4.14.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0
flask==3.0.0
flask-sqlalchemy==3.1.1